Separate out new code into separate modules
