+++ crates/eww/src/state/mod.rs
@@ -1,0 +1,6 @@
+mod inheritance_map;
+pub mod scope;
+pub mod scope_graph;
+
+#[cfg(test)]
+mod test;
